requirements.txt
python-dotenv
requests
reportlab
pillow
